Changed the direction of the treat from left to above by changing fly.x to fly.y
in resetfly I swaped the numbers making x the random value and y to -2 so it looks like the treats are being tossed from somewhere out of the frame.
I could make it a random number to make it look like they were being tossed by the person standing in the slippers. (Me...hehe)