Used a variable so I can make the treats move faster as the game progresses!
Sat a variable (let difficulty = 1;) so I can use it in the moveFly function, to make the treats mover faster. And then I added it to 'function draw'  as well so they will move faster over the span of the entire game, instead of them moving faster individually pr. frame.